refactor(altium): split reader.rs and primitives.rs into submodules
Completes the module split flagged in #105 (pcblib/mod.rs was split in #109).
Behaviour-neutral — purely navigability, no logic changes.

reader.rs (2489 lines) -> reader/{mod,parsers,models}.rs:
  - mod.rs: byte/block helpers, wide-strings, unique-ids, stream dispatch, tests
  - parsers.rs: per-primitive binary parsers (pad/via/track/arc/text/region/
    fill/component-body)
  - models.rs: embedded 3D-model stream parsing + zlib decompression

primitives.rs (1580 lines) -> primitives/{mod,pads,shapes,text,models3d,layer}.rs:
  - grouped by primitive family; mod.rs re-exports preserve `primitives::X` paths

Closes #105
